cps-enterprise-dcs/local-agent/requirements.txt
# CP'S Enterprise DCS - Local Agent Dependencies
# Python 3.11+

# Core Framework
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Cryptography & Security
cryptography>=42.0.0
hashlib2>=1.1.0
PyJWT>=2.8.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Event Store & Messaging
aiokafka>=0.9.0
redis>=5.0.0
hiredis>=2.3.0

# Async & Concurrency
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
orjson>=3.9.0

# Machine Learning (Edge)
scikit-learn>=1.3.0
onnxruntime>=1.16.0

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-grpc>=0.42b0

# Configuration & Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=23.2.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
